src/projector/requirements.txt
# TensorFlow & TensorBoard (for checkpoint generation)
tensorflow==2.18.0
tensorboard==2.18.0

# NumPy (required by TensorFlow)
numpy>=1.24.0,<2.0.0

# NATS (message queue)
nats-py>=2.10.0

# Async support
aiohttp>=3.11.0

# Local dependencies (installed via COPY in Dockerfile)
# echomind_lib (local)
